Reorganizing app architecture by creating folders per each component
