+++ src/Components/Form/FormTask.tsx
@@ -1,11 +1,11 @@
 import { useState, useContext, FormEvent } from "react";
 import { useNavigate } from "react-router-dom";
 import nextId from "react-id-generator";
-import { StyledFormTask, FormBody } from "./Styles/FormTask.styled";
-import { StyledButton } from "./Styles/Buttons.styled";
-import ThemeContext from "../Contexts/ThemeContext";
-import { ITask } from "../Types/index";
-import { createRequest } from "./API Requests/Requests";
+import { StyledFormTask, FormBody } from "./FormTask.styled";
+import { StyledButton } from "../Styles/Buttons.styled";
+import ThemeContext from "../../Contexts/ThemeContext";
+import { ITask } from "../../Types/index";
+import { createRequest } from "../API Requests/Requests";
 
 const FormTask = () => {
   const [inputTitle, setInputTitle] = useState("");
