+++ src/Components/Themes/ColorTheme.tsx
@@ -1,6 +1,6 @@
 import React, { useContext, MouseEvent } from "react";
-import ThemeContext from "../Contexts/ThemeContext";
-import { StyledColorTheme } from "./Styles/ColorTheme.styled";
+import ThemeContext from "../../Contexts/ThemeContext";
+import { StyledColorTheme } from "./ColorTheme.styled";
 
 const ColorTheme = () => {
   const [themes, setThemes] = useContext(ThemeContext);
