+++ src/index.html
@@ -17,6 +17,6 @@
   <body>
     <div id="modal"></div>
     <div role="main" id="root">not rendered</div>
-    <script type="module" src="./App.tsx"></script>
+    <script type="module" src="./Components/App/App.tsx"></script>
   </body>
 </html>
